with modify and delete buttons working with alert

+++ my-express-app/client/src/App.js
@@ -6,13 +6,15 @@ import PatientFile from "./Components/PatientFile";
 import HomeView from "./Components/HomeView";
 import ExercisesView from "./Components/ExercisesView";
 import UrlToShare from "./Components/UrlToShare";
-import ExercisesList from "./Components/ExercisesList";
-
+import { useNavigate } from "react-router-dom";
 
 function App() {
   
   const [patients, setPatients] = useState([]);
   //when the page loads, fetch the data from the server
+
+  const navigate = useNavigate(); //to go to the search page if I delete a patient
+
   useEffect(() => {
     getPatients();
   }, []);
@@ -32,11 +34,16 @@ async function getPatients() {
 }
 }
 
-//NATALIA DID THESE 2 FUNCTIONS BELOW TO DELETE AND MODIFY PATIENTS
 //Delete a patient
-async function deletePat(id) {
+async function deletePatient(id) {
   // Define fetch() options
-  let options = {
+  let confirm = window.confirm("Are you sure you want to delete this patient? This patient's programs and exercises will be deleted too.")
+  console.log(confirm);
+  
+  if (confirm) {
+    navigate("/");
+
+    let options = {
       method: 'DELETE'
   };
 
@@ -52,21 +59,21 @@ async function deletePat(id) {
       console.log(`Server error: ${err.message}`);
   }
 }
+  }
 
-//PUT Modify patient's data
-async function modifyPat(id) {
-  let patient = patients.find(p => p.id === id);
+// PUT: Modify patient inputs 
+async function modifyPatient(id, formData) {
 
   let options = {
       method: 'PUT',
       headers: { 'Content-Type': 'application/json' },
-      body: JSON.stringify(patient)
+      body: JSON.stringify(formData)
   };
 
   try {
       let response = await fetch(`/patients/${id}`, options);
       if (response.ok) {
-          let patients= await response.json();
+          let patients = await response.json();
           setPatients(patients);
       } else {
           console.log(`Server error: ${response.status} ${response.statusText}`);
@@ -78,17 +85,15 @@ async function modifyPat(id) {
 
   return (
     <div className="bg-info">
-    
-  
+      
       <Routes className="center">
         <Route path="/" element= {<HomeView patients={patients}/>}  />
-        <Route path="/patients/:patientId" element={<PatientFile patients={patients}/>} />
+        <Route path="/patients/:patientId" element={<PatientFile modifyPatient={modifyPatient} deletePatient={deletePatient} patients={patients}/>} />
         <Route path="/programs/program/:programId" element={<ExercisesView />} />
         <Route path="/exercises/:programId" element={<UrlToShare />} />
       </Routes>
-      
-    
-    </div>
+
+    </div> 
   );
 }
